Original file.
I’m dating this commit December 4, 2013 because this is the last
modification time of the /usr/local/bin/prompt file on my current
computer, but the file is much older. I’m pretty sure I wrote it
while being a student at the ENS Lyon, which would mean somewhere
between 2005 and 2008.